upgrade to new web component standards
