+++ tsconfig.json
@@ -3,12 +3,11 @@
     "module": "es2020",
     "target": "es2017",
     "strict": true,
+    "moduleResolution": "node",
     "declaration": true,
     "outDir": "dist",
-    "removeComments": true,
-    "preserveConstEnums": true
+    "lib": ["dom", "dom.iterable", "es2020"],
+    "removeComments": true
   },
-  "files": [
-    "src/index.ts"
-  ]
+  "files": ["src/index.ts"]
 }
